+++ src/types/adapter.d.ts
@@ -3,7 +3,6 @@
 import { ACSAdapterLogger } from "../utils/loggers";
 import AMSFileManager from "../external/ACSAdapter/AMSFileManager";
 import { ChatClient } from "@azure/communication-chat";
-import PostChatContext from "../core/PostChatContext";
 
 type AdapterEgressIngressFunction = (arg: any) => Promise<any>;
 type GetStateFunction = (name: string) => unknown;
@@ -49,23 +48,3 @@ export type AdapterCreator = (
 	chatClient: ChatClient,
 	logger: ACSAdapterLogger,
 	featureOption: IAdapterInitFeatureOption) => IChatAdapter;
-
-export type OmnichannelGenericResponse = {
-	data : {} | undefined;
-	error? : string;
-	success: boolean;
-}
-
-export type GetPrechatSurveyResponse = OmnichannelGenericResponse;
-export type GetDataMaskingRulesResponse = OmnichannelGenericResponse;
-export type EmailLiveChatTranscriptResponse = OmnichannelGenericResponse;
-export type GetLiveChatTranscriptResponse = OmnichannelGenericResponse;
-
-export type GetPostChatSurveyContextResponse = {
-	data : PostChatContext | undefined;
-	error? : string;
-	success: boolean;
-};
-
-export type GetAgentAvailabilityResponse = OmnichannelGenericResponse;
-export type CreateChatAdapterResponse = OmnichannelGenericResponse;
